+++ packages/bluesky/lib/src/entities/embed_record_with_media.dart
@@ -7,7 +7,8 @@
 // 📦 Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'embed_images.dart';
+import 'converter/embed_media_converter.dart';
+import 'embed_media.dart';
 import 'embed_record.dart';
 
 part 'embed_record_with_media.freezed.dart';
@@ -20,7 +21,7 @@ class EmbedRecordWithMedia with _$EmbedRecordWithMedia {
     @JsonKey(name: '\$type')
         String type,
     required EmbedRecord record,
-    required EmbedImages media,
+    @EmbedMediaConverter() required EmbedMedia media,
   }) = _EmbedRecordWithMedia;
 
   factory EmbedRecordWithMedia.fromJson(Map<String, Object?> json) =>