+++ packages/bluesky/lib/src/entities/embed_record_with_media.g.dart
@@ -18,8 +18,10 @@ _$_EmbedRecordWithMedia _$$_EmbedRecordWithMediaFromJson(Map json) =>
               (v) => v as String? ?? 'app.bsky.embed.recordWithMedia'),
           record: $checkedConvert('record',
               (v) => EmbedRecord.fromJson(Map<String, Object?>.from(v as Map))),
-          media: $checkedConvert('media',
-              (v) => EmbedImages.fromJson(Map<String, Object?>.from(v as Map))),
+          media: $checkedConvert(
+              'media',
+              (v) => const EmbedMediaConverter()
+                  .fromJson(v as Map<String, dynamic>)),
         );
         return val;
       },
@@ -31,5 +33,5 @@ Map<String, dynamic> _$$_EmbedRecordWithMediaToJson(
     <String, dynamic>{
       r'$type': instance.type,
       'record': instance.record.toJson(),
-      'media': instance.media.toJson(),
+      'media': const EmbedMediaConverter().toJson(instance.media),
     };