+++ packages/bluesky/lib/src/entities/embed_view_images.dart
@@ -4,7 +4,6 @@
 
 // ignore_for_file: invalid_annotation_target
 
-// 📦 Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 import 'embed_view_images_view.dart';