+++ packages/bluesky/lib/src/entities/embed_view_record_with_media.dart
@@ -7,7 +7,8 @@
 // 📦 Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'embed_view_images.dart';
+import 'converter/embed_view_media_converter.dart';
+import 'embed_view_media.dart';
 import 'embed_view_record.dart';
 
 part 'embed_view_record_with_media.freezed.dart';
@@ -18,7 +19,7 @@ class EmbedViewRecordWithMedia with _$EmbedViewRecordWithMedia {
   const factory EmbedViewRecordWithMedia({
     @JsonKey(name: '\$type') required String type,
     required EmbedViewRecord record,
-    required EmbedViewImages media,
+    @EmbedViewMediaConverter() required EmbedViewMedia media,
   }) = _EmbedViewRecordWithMedia;
 
   factory EmbedViewRecordWithMedia.fromJson(Map<String, Object?> json) =>