+++ packages/bluesky/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.8
 
 - Added `protocol` option. Defaults to `HTTPS`. ([#269](https://github.com/myConsciousness/atproto.dart/issues/269))
+- Fixed bugs of `media` field in `EmbedRecordWithMedia` and `EmbedViewRecordWithMedia`. ([#282](https://github.com/myConsciousness/atproto.dart/issues/282))
 
 ## v0.3.7